[Var] ancient: Update to commit c79987baa02802999a80af2a2bc36c3f01d07825 (2023-07-11).
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@19460 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ include/ancient/src/BLZWDecompressor.hpp
@@ -26,7 +26,7 @@ private:
 	const Buffer	&_packedData;
 
 	uint32_t	_maxBits=0;
-	size_t		_stackLength=0;
+	uint32_t	_stackLength=0;
 };
 
 }
